refactor: #501 Todo 기본 간격 ㄱ밧 변경 및 sort 값에 음수 허용하도록 변경 Todo의 기본 간격 값을 3072로 증가시켰다. Todo의 Sort값이 음수가 되도 허용하도록 변경하였다.

+++ backend/src/main/java/site/coduo/todo/domain/Todo.java
@@ -19,7 +19,7 @@ public class Todo {
             final Long id,
             final PairRoom pairRoom,
             final String content,
-            final int sort,
+            final double sort,
             final boolean isChecked
     ) {
         this(
